refactor: 알림 타입 위치 변경 적용 (#236)
